Create utils folder. Move utils.py into the util folder. Create utils/read_files.py. Move relevant functions from utils.py to utils/read_files.py
