+++ statistics/utils/read_files.py
@@ -1,11 +1,7 @@
 import os
 import re
-import argparse
 
-import numpy as np
 import pandas as pd
-
-from scipy.stats import spearmanr
 
 DICT_DISC_LABELS = {
                     'C1/C2': 2,
@@ -15,15 +11,6 @@ DICT_DISC_LABELS = {
                     'C5/C6': 6,
                     'C6/C7': 7
                     }
-
-
-class SmartFormatter(argparse.HelpFormatter):
-
-    def _split_lines(self, text, width):
-        if text.startswith('R|'):
-            return text[2:].splitlines()
-        # this is the RawTextHelpFormatter._split_lines
-        return argparse.HelpFormatter._split_lines(self, text, width)
 
 
 def fetch_participant_id(filename):
@@ -233,42 +220,3 @@ def read_electrophysiology_anatomical_and_motion_file(file_path, df_participants
     electrophysiology_df = electrophysiology_df.set_index(['participant_id'])
 
     return anatomical_df, motion_df, electrophysiology_df
-
-
-def format_pvalue(p_value, alpha=0.001, decimal_places=3, include_space=False, include_equal=True):
-    """
-    Format p-value.
-    If the p-value is lower than alpha, format it to "<0.001", otherwise, round it to three decimals
-    :param p_value: input p-value as a float
-    :param alpha: significance level
-    :param decimal_places: number of decimal places the p-value will be rounded
-    :param include_space: include space or not (e.g., ' = 0.06')
-    :param include_equal: include equal sign ('=') to the p-value (e.g., '=0.06') or not (e.g., '0.06')
-    :return: p_value: the formatted p-value (e.g., '<0.05') as a str
-    """
-    if include_space:
-        space = ' '
-    else:
-        space = ''
-
-    # If the p-value is lower than alpha, return '<alpha' (e.g., <0.001)
-    if p_value < alpha:
-        p_value = space + "<" + space + str(alpha)
-    # If the p-value is greater than alpha, round it number of decimals specified by decimal_places
-    else:
-        if include_equal:
-            p_value = space + '=' + space + str(round(p_value, decimal_places))
-        else:
-            p_value = space + str(round(p_value, decimal_places))
-
-    return p_value
-
-
-def compute_spearmans(a, b):
-    """
-    Compute Spearman's correlation coefficient and p-value for two arrays.
-    """
-    a = np.array(a)
-    b = np.array(b)
-    return spearmanr(a, b)
-
